Added test cases

+++ src/main/java/com/nagoorkhan/productservice/model/response/ProductResponseVO.java
@@ -1,11 +1,12 @@
 package com.nagoorkhan.productservice.model.response;
 
-import com.nagoorkhan.productservice.model.business.ProductVO;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import org.springframework.data.annotation.Id;
+
+import java.math.BigDecimal;
 
 @NoArgsConstructor
 @AllArgsConstructor
@@ -15,6 +16,6 @@ public class ProductResponseVO {
     @Id
     private String pid;
     private String productName;
-    private Long price;
+    private BigDecimal price;
     private String description;
 }
